refactor(web): remove `ConfigParser` from iron-remote-desktop interface
