--- crates/iron-remote-desktop/src/config.rs
@@ -1,10 +1,0 @@
-pub trait ConfigParser {
-    #[must_use]
-    fn create(config: &str) -> Self;
-
-    #[must_use]
-    fn get_str(&self, key: &str) -> Option<String>;
-
-    #[must_use]
-    fn get_int(&self, key: &str) -> Option<i32>;
-}
