--- web-client/iron-remote-desktop/src/interfaces/ConfigParser.ts
@@ -1,4 +1,0 @@
-export interface ConfigParser {
-    get_str(key: string): string | null;
-    get_int(key: string): number | null;
-}
